@@ -18,30 +18,28 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+ sdk: ">=2.11.0 <3.0.0"
 
 dependencies:
-  ntp: ^1.0.7
-  video_thumbnail: ^0.2.5+1
-  video_player: ^0.9.0
-  firebase_core: ^0.4.0+9
-  firebase_analytics: ^5.0.6
-  firebase_storage: ^3.1.6
-  image_picker: ^0.6.7+11
-  camera: 0.5.8+17
-  path_provider: ^1.6.27
-  path: ^1.7.0
-  firebase_auth: ^0.16.0
-  adobe_xd: ^1.0.0+1
-  flutter_launcher_icons: ^0.8.1
-  cloud_firestore: ^0.13.7
-  http: ^0.12.2
-  provider: ^4.3.2+3
-  firebase_messaging: ^6.0.1
-  visibility_detector: ^0.1.5
-  scrollable_positioned_list: ^0.1.10
   flutter:
     sdk: flutter
+  ntp: ^2.0.0
+  video_thumbnail: ^0.3.3
+  video_player: ^2.1.6
+  firebase_core: ^1.3.0
+  firebase_storage: ^8.1.3
+  image_picker: ^0.8.0+3
+  camera: ^0.8.1+3
+  path_provider: ^2.0.2
+  path: ^1.8.0
+  firebase_auth: ^1.4.1
+  adobe_xd: ^2.0.0+1
+  flutter_launcher_icons: ^0.9.0
+  cloud_firestore: ^2.2.2
+  http: ^0.13.3
+  provider: ^5.0.0
+  visibility_detector: ^0.2.0
+  scrollable_positioned_list: ^0.1.10
 
 
   # The following adds the Cupertino Icons font to your application.
